Add an extra block in the search scripts to load extra scripts
Use case:
I want to change the search template to use directives instead of
changing angular in the template. But I need the module to be loaded
first to add the directive to the scope.
If I use extra_script it won't work as expected
1. I need to load it in the index template and therefore be available
everywhere
2. if I do that the search scripts are not loaded yet

This new block makes it easier to extent the search with more
directives, scripts, etc